azure-insurance-agentcore/requirements.txt
fastapi
uvicorn
pydantic
openai
azure-search-documents
azure-cosmos
azure-core
azure-identity